feat(core): split createRuntimeExtras into a neutral brand and a react binding
